spring mvc

+++ src/main/src/main/java/springdata/repository/CustomerRepo.java
@@ -1,6 +1,7 @@
 package springdata.repository;
 
 import entities.Customer;
+import org.springframework.cache.annotation.Cacheable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -9,5 +10,5 @@ import java.util.List;
 
 @Repository
 public interface CustomerRepo extends JpaRepository<Customer, BigDecimal> {
-    List<Customer> findByCreditLimitBetween(int minLimit, int maxLimit);
+    List<Customer> findByCreditLimitBetween(BigDecimal minLimit, BigDecimal maxLimit);
 }
